# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MTC App Api",
        "description": "MTC Application OpenApi Documentation",
        "contact": {
            "email": "aminetemadifar.w@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "localhost"
        }
    ],
    "paths": {
...
        "/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get all lessons",
                "description": "Retrieve a list of all lessons",
                "operationId": "e1abe40dbd1870d06943517976c11507",
                "responses": {
                    "200": {
                        "description": "List of lessons retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LessonResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create a new lesson",
                "description": "Create a new lesson with the provided data",
                "operationId": "b196bee153377a601911ff0db5a9aa01",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "lesson_type_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Physics"
                                    },
                                    "lesson_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "theory_unit": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "theory_time": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "practical_unit": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "practical_time": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LessonResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lessons/{lesson}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get a specific lesson",
                "description": "Retrieve details of a specific lesson",
                "operationId": "bf10e0dab082ab821b34120410bcc4bd",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LessonResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lesson not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update a lesson",
                "description": "Update the details of a specific lesson",
                "operationId": "9f9c81867de31574b33e1add9b82416b",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Advanced Physics"
                                    },
                                    "lesson_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "theory_unit": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "theory_time": {
                                        "type": "integer",
                                        "example": 60
                                    },
                                    "practical_unit": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "practical_time": {
                                        "type": "integer",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LessonResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lesson not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete a lesson",
                "description": "Delete a specific lesson",
                "operationId": "49e900f2761729d66df94cab08a95d75",
                "parameters": [
                    {
                        "name": "lesson",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Lesson not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lesson not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LessonResource": {
                "title": "Lesson Resource",
                "description": "Lesson resource schema",
                "properties": {
                    "id": {
                        "description": "Lesson code",
                        "type": "integer",
                        "example": 1
                    },
                    "lesson_type": {
                        "$ref": "#/components/schemas/LessonTypeResource"
                    },
                    "title": {
                        "description": "Lesson name",
                        "type": "string",
                        "example": "Mathematics"
                    },
                    "theory_unit": {
                        "description": "Theory unit count",
                        "type": "integer",
                        "example": 2
                    },
                    "theory_time": {
                        "description": "Theory time in minutes",
                        "type": "integer",
                        "example": 45
                    },
                    "practical_unit": {
                        "description": "Practical unit count",
                        "type": "integer",
                        "example": 1
                    },
                    "practical_time": {
                        "description": "Practical time in minutes",
                        "type": "integer",
                        "example": 30
                    },
                    "status": {
                        "description": "Lesson status",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "LessonTypeResource": {
                "title": "Lesson Type Resource",
                "description": "Lesson type resource schema",
                "properties": {
                    "id": {
                        "description": "Lesson type code",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Lesson type name",
                        "type": "string",
                        "example": "Theory"
                    },
                    "status": {
                        "description": "Lesson type status",
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}